+++ src/nvim/eval/vars.c
@@ -672,7 +672,7 @@ static char *ex_let_one(char *arg, typval_T *const tv, const bool copy, const bo
 
       if (!failed) {
         if (opt_type != gov_string || s != NULL) {
-          char *err = set_option_value(arg, n, s, opt_flags);
+          char *err = set_option_value(arg, (long)n, s, opt_flags);
           arg_end = p;
           if (err != NULL) {
             emsg(_(err));
